First code checkin, verified 100% by running the app and checking its output
This one was a bit tricky as it depends on two outside libraries, but as I forked both of them with the intent of working on issues listed at their Git sites, I was able to slightly modify their Eclipse and Maven files to specify Java1.8 vs. Java 1.6 or Java 11. There was also a mistake in the POM regarding the Group ID and Artifact ID, and the version of a Maven plug-in was too recent so had to be backed down as well. This means the project dependency is on my fork vs. the main root for Orson PDF and JFreeSVG. This issue may go away once I upgrade to Java 14.