Removed Nimbus as the startup LAF due to what was learned yesterday from a key developer
Although at one point there may have been a reason to default to Nimbus during some instability in Windows LAF and Aqua LAF (and I no longer remember how far back this was, as I've mostly been doing JavaFX since 2012, starting in hybrid mode and going pure JFX by 2015 to 2016), I was assured yesterday by a key developer that it is not advisable to have a toolkit default to Nimbus LAF and that the native look of the default LAF for each platform is quite consistent and stable these days. So I have removed the LAF setter from the startup code in the demo frame.